+++ jdevelops-authentication/jdevelops-authentication-sas/jdevelops-authentication-sas-server/src/main/java/cn/jdevelops/authentication/sas/server/core/config/SasProperties.java
@@ -37,6 +37,9 @@ public class SasProperties {
     }
 
     public SasAuthorizeHttpRequests getRequests() {
+        if(null == requests){
+            return new SasAuthorizeHttpRequests();
+        }
         return requests;
     }
