+++ jdevelops-authentication/jdevelops-authentication-sas/jdevelops-authentication-sas-server/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,10 @@
-cn.jdevelops.authentication.sas.server.ImportConfig
+cn.jdevelops.authentication.sas.server.core.config.SasProperties
+cn.jdevelops.authentication.sas.server.core.config.ImportConfig
+cn.jdevelops.authentication.sas.server.core.config.AuthorizationServerConfig
+cn.jdevelops.authentication.sas.server.core.config.DefaultSecurityConfig
+cn.jdevelops.authentication.sas.server.core.controller.ServerController
+cn.jdevelops.authentication.sas.server.core.controller.AuthorizationConsentController
+
+
+
